Move `generateLibVersion` task to `kotlin-sdk-core` and update build configuration
