+++ buildSrc/src/main/kotlin/mcp.multiplatform.gradle.kts
@@ -10,25 +10,6 @@ plugins {
     id("org.jetbrains.kotlinx.atomicfu")
 }
 
-// Generation library versions
-val generateLibVersion by tasks.registering {
-    val outputDir = layout.buildDirectory.dir("generated-sources/libVersion")
-    outputs.dir(outputDir)
-
-    doLast {
-        val sourceFile = outputDir.get().file("io/modelcontextprotocol/kotlin/sdk/LibVersion.kt").asFile
-        sourceFile.parentFile.mkdirs()
-        sourceFile.writeText(
-            """
-            package io.modelcontextprotocol.kotlin.sdk
-
-            public const val LIB_VERSION: String = "${project.version}"
-            
-            """.trimIndent()
-        )
-    }
-}
-
 kotlin {
     jvm {
         compilerOptions.jvmTarget = JvmTarget.JVM_1_8
@@ -41,10 +22,4 @@ kotlin {
 
     explicitApi = ExplicitApiMode.Strict
     jvmToolchain(21)
-
-    sourceSets {
-        commonMain {
-            kotlin.srcDir(generateLibVersion)
-        }
-    }
 }
